FIT-21-details-food: fix structure
